Flesh out iiglue reader to read and store inferred array info.
We are now loading JSON-formatted iiglue analysis results and associating those results with the appropriate LLVM objects representing program components.  However, the reader is customized for just the iiglue information we actually need.  Right now, we only care which formal function arguments are arrays.  So we extract that and provide a convenient API for querying it from other passes.  Any other iiglue analysis output is quietly ignored.