Create some simple unit tests using Google Mock and Google Test.
Unfortunately, I don't think we're actually going to use this stuff.  LLVM is quite selective in making class methods virtual.  Mocking non-virtual methods requires extensive use of templates in application code, which harms readability and increases build time.  So I don't plan to merge this experiment into master at this point.  Rather, I'm saving this branch for historical reference, since it took a moderate amount of effort to get up and running.  If we decide to take a second look at Google Mock or Google Test in the future, perhaps this will be a good starting point.