Define a typedef alias for a complex type that appears in a public API.
I don't necessarily define typedef aliases for all nontrivial template
instantiations.  But when a tricky type appears in the public API of a
class, you probably will be using that type more than once: in the method
declaration, in the method definition, and in most or all callers.  Once
a complex template instantiation is being used that frequently, giving it
a shorter name is helpful.