Use a handy boost feature for iterating over numeric intervals.
This lets us iterate over that interval using the "for (:)" style of loop,
which I find more readable, less error-prone, and potentially even faster.
For example, it's faster in this case because we compute the loop upper
bound just once instead of calling "call->getNumArgOperands()" on every
iteration.

IMHO, with C++11 and Boost available, there are very few scenarios left
where an old-style "for (;;)" loop is still appropriate at all.