Visually associate "*" with the symbol being declared, not with the type.
This is just a white space tweak, but it puts one in the right mindset to
understand more complex, multi-variable declarations like "int *x, y".
That declares "x" as a pointer to int, but "y" as just an int.  If you
space that out as "int* x, y" it gives a strong (but wrong) impression that
both "x" and "y" are being declared as pointers to int.