Reduce copying of containers when inserting them into enclosing containers.
The general strategy being used here is to reserve a slot in the enclosing
container *first*.  That slot will initially be filled with an empty,
newly-constructed instance of the inner container.  We hold a reference to
that inner container, and then subsequently modify it directly.  That way
we don't need to copy the inner container to insert it into the enclosing
container later, since it's already in place.