Collect each function's call sites just once, not once per array argument.
The "firstTime" flag was doing the right thing by preventing us from doing
this work after the first iteration.  But we were still doing it once per
array argument.  There's no need for that, since the collected set of call
sites does not depend on which argument we are currently looking at.

By the way, there's some cute use of boost::lambda::_1 and
unordered_set::emplace in here when we collect the call instructions and
then build a set out of them.  You may learn a few new C++11 and Boost
tricks by looking at this code!