Tighter code for checking whether any map entry has a desired property.
The original code here was quite wasteful.  It was iterating over all
(key, value) pairs in the map, ignoring the value, then using the key
to look up the value in the map.  Sometimes you really do need all
(key, value) pairs.  But in this particular case, we don't care about
the keys at all.  It's sufficient to just iterate over the values.