Refactor lots of common scons code used in test subdirectories.
Discovery and running of tests is now quite automated.  That is
convenient but fairly inflexible.  I've left little room for
variation: only the things that already vary from one subdirectory to
another are accommodated here.  If we decide that more flexibility is
needed, we can certainly add that back in later.