Update where we look for array annotations in `iiglue` output.
`iiglue`'s JSON output recently changed its structure slightly.
Annotations for each function parameter are stored under a
"parameterAnnotations" key.  Previously the value of that key looked
like:

```
	[
	  {"contents": [], "tag": "PANotNull"},
	  {"contents": 1, "tag": "PAArray"}
	]
```

The new output still uses a "parameterAnnotations" key, but now the
value of that key looks like:

```
	[
	  {"PANotNull": []},
	  {"PAArray": 1}
	]
```

This change appears to be due to an update to the Haskell module
`iiglue` uses to serialize data structures as JSON.  I have not
attempted to remain backward-compatible with the old style; you must
use `iiglue` output in the new style only.