Use argument slot numbers if we cannot find a formal argument's name.
Formals can be unnamed if the code was really written that way; this is a
common style in C++ for formal arguments that are not actually used by the
called function.  Formals can also be unnamed if the function is declared
but not defined in this module.  Either way, I find it better to see an
argument described as "foo::2" (for argument #2 to function foo) rather
than as just "foo::" if the name is unknown.