Allow more than one iiglue file among the list of sources.
We now consider the source list to be any jumbled mix of LLVM bitcode
files, LLVM plugins, and iiglue analysis results.  To produce the right
command line for running our own plugin, we walk through the source list
and emit whatever flag(s) should correspond to each source file depending
on its extension.  For example, a file ending in ".json" is taken to be
an iiglue analysis results file, and is preceded on the command line with
an "-iiglue-read-file" flag.