Add a test that reads multiple iiglue analysis result files.
The only thing a little weird about this is that we don't get names for the
formal arguments to externally-defined functions.  So in the IIGlueReader
debug output, these arguments are identified by slot number (counting from
zero) rather than by name.  That's about the best we can do with the
information Clang and LLVM provide.  It's not as readable as having proper
formal argument names, but it's still deterministic and unambiguous, and
therefore suitable for use in a regression test.