Simplify traversePHIs() API by offering an alternate entry point.
Previously the caller had to declare an unordered set of phi nodes
found so far, and pass that in as a by-reference argument.  But
really, the caller never cares about that set.  It's an implementation
detail that prevents infinite loops when traversePHIs() recursively
calls itself.

The simpler API is an alternate entry point that omits that set of
already-found phi nodes.  This alternate entry point takes care of
declaring that set and calling into the main traversePHIs() logic.
This makes life easier for the caller by letting the caller remain
ignorant of this phi node set, which is really an implementation
detail.  It also means that set will go out of scope sooner, thereby
releasing its memory sooner.