Prefer iteration over ranges rather than old-fashioned "for" loops.
This could be considered a matter of personal taste.  I like the irange()
approach better because it's a higher level of abstraction that more
clearly documents our intent.  Using this style also means we don't have to
compute and stash the ending value ourselves.  Istead, the endinf value is
just passed in as an argument to irange(), and the range itself is
responsible for knowing when to stop.