Avoid two map lookups when inserting a value if not already already there.
Previously we were checking for the value already being present, and then
doing an insertion if it was not already there.  That requires two map
lookups.  Instead, we can always just perform the insertion.  This is not
a multimap, so we won't end up with duplicates.  Furthermore, the insert
method's return value will tell us whether the value was already present
or not.  So this gives us the same overall effect but with just one map
lookup instead of two.