Verify size-matched containers just once, rather than on every iteration.
When walking through two containers that should have the same size, it's
certainly proper to make sure that neither will run out of elements early.
That must be done on each loop iteration if the containers' sizes are not
available right at the start.  But if we do know the containers' sizes
before the loop even begins, then we can check their sizes just once
instead.