Extract enum Answer into its own file, and add some Boost ptree support.
Specifically, the ptree support takes the form of teaching Boost how to
convert Answer values to and from strings.  Converting to string is useful
when writing a ptree into a file, and converting from string is useful when
reading a ptree from a file.  Nothing is using either of those converstions
right now, but soon I'll be changing our JSON file format so that Answer
values are stored as informative strings (e.g., "don't care") rather than
as opaque integer values (e.g., 0).