Restructure JSON file format, and use Boost's JSON output routines.
Previously the file format consisted of a top-level dictionary with
one "library_functions" key.  The value for that key was an array of
per-function dictionaries, and one of the keys for each per-function
dictionary was "name", whose value was the function name.  For
example:

    {
        "library_functions": [
            {
	        "name": "someFunction",
		...
	    },
            {
	        "name": "anotherFunction",
		...
	    },
	    ...
	]
    }

The new structure still has the same top-level dictionary with just a
"library_functions" key.  However, the value at that key is now itself
a dictionary (not an array).  Each key in this dictionary is a
function name, and the value at that key is the information for that
function.  For example:

    {
        "library_functions": {
            "someFunction": {
		...
	    },
	    "anotherFunction": {
		...
	    },
	    ...
        }
    }

So in brief, function names now serve as unique keys to look up
per-function information in the "library_functions" dictionary.

Along with this, analysis answers (values of type "enum Answer") are
now stored as descriptive strings rather than small integers.  I find
"don't care" much more informative than "0".

One odd side effect of this change is that Boolean values are written
out as quoted strings: "true" instead of just true.  This is an
unfortunate limitation of Boost property trees, which internally
convert everything to strings rather than preserving values' original
types.  This makes no difference when reading JSON, and in fact "true"
and true will both be handled just fine.  I'd prefer that Booleans be
written out without the double quotes, but that's not currently
possible if we want to use Boost's output routines.

Revamp the analysis output routines to use Boost property trees, just
as we were already doing for analysis input.  Instead of building up
long strings of JSON-formatted text, we now just populate a Boost
property tree and use Boost's own JSON output routine to store it into
a file.

Revamp the "NullAnnotatorSummary.py" script to match the changes made
to "NullAnnotator.cc".  I also took this opportunity to use some more
efficient alternatives in the Python code.  For example, I've switched
one variable from a list to a set: this variable was not
order-sensitive, and we were *always* checking whether it already
contained any item we wanted to add.  Many print statements have
changed from using string concatenation to either comma-delimited
arguments or "%"-style format strings.  Instead of just asserting that
two dictionaries have the same length, assert that they have identical
keys: a much stronger property.

Unfortunately I have not been able to fully test these big changes.
As far as I can tell, we have no regression tests that *read* any
saved NullAnnotator analysis outputs, and also none that run the
"NullAnnotatorSummary.py" script.  So it's quite possible that
regressions have crept in.  I also have not updated the "*.json"
analysis output files already present in the repository.  I'm not
entirely convinced that these belong in the repository at all; Alisa
and I should discuss this further.

Also in this change, all BOOST_FOREACH loops have been replaced with
standard C++11 "for" loops.  There's really no longer any good reason
to use BOOST_FOREACH now that C++11 has become mainstream.