Extract common code from two different traversePHIs() implementations.
We now have a nice encapsulation of searching backward from a starting
value to arguments, crossing zero or more phi nodes in between.  Subclasses
provide their own handlers to be called on each reached argument.  Right
now we use two such subclasses.  One collects all the reached arguments in
a set.  The other stops as soon as it finds a single, specific argument of
interest.