Avoid visiting any nodes twice.
Previously we were only avoiding redundant visits to phi nodes.  That
was enough to keep us out of infinite loops, but we could still visit
other kinds of Value nodes multiple times if they were operands to
multiple different phi nodes.  This change closes that loophole.  I'm
pleased that this was so self-contained: hooray for encapsulation!