Simplify container use slightly.
The high-level goal here is to iterate through a map's keys in sorted
order.  The underlying map, however, is a std::unordered_map.  So
we'll need to do some sorting somewhere.  Previously we were building
a sorted set of map keys, then looking up each key in the original map
to get that key's value.  Now we build a sorted map, which means we
don't need to do a second lookup to find the corresponding data value.