Suppress output about arguments that have no sentintel checks.
When running in "-overreport" mode, we examine many more arguments, most of
which are probably not even arrays at all.  This change lets us avoid
lots of extra output noise for these uninteresting arguments.

Unfortunately, this does also remove output for interesting array arguments
that just happen to not be sentinel-checked.  That might have been useful
when debugging to distinguish them from arguments that were not examined at
all.  If we eventually find that this is a problem, we can add an option of
some kind to bring this extra output back.