Add LLVM_SRA scons command-line option.
Previously we were hard-coded to look for the llvm-sra build tree in a
specific subdiretory within Alisa's home directory.  Now we can put the
llvm-sra build tree wherever we want and tell scons where to find it.
The default location is in the tools area of "/p/polyglot", where we
hope to eventually have a good, working SRA build.