Work around some shell confusion.
Running "sage -sh" is supposed to run your own preferred shell augmented
with various SAGE-related environment variables.  Unfortunately, this
doesn't work correctly for my preferred shell (fish).  So here we
override the $SHELL environment variable to set it to something that
will work correctly.  This is all for non-iterative use anyway, so the
actual shell used doesn't really matter, as long as it works.