Decouple llvm-config and llvm-sra locations.
I'm in the process of moving us to a build strategy where llvm-sra does
not have its own complete llvm tree, but rather relies on one already
built elsewhere.  CArrayIntrospection then uses that same llvm tree,
with the path to llvm-config being the means by which the llvm tree is
found and probed.