Simplify plugin loading.
Now we link our plugin directly against the llvm-sra plugin, which in
turn links against the llvm-sage plugin, which in turn links against the
llvm-python plugin.  So all we need to do is "-load" our own plugin.
The others will be pulled in automatically.  That's much nicer than
having to "-load" all four plugins ourselves, in just the right order.