Extract and generalize name-based comparator.
It's sometimes useful to sort things by name when printing, especially
for regression testing.  Previously we had *two* copies of a
BasicBlockCompare class that provided this sorting comparison.  Why did
we have two?  Sloppy copy and paste, no doubt.

Now we have a single NameCompare class that provides this functionality.
Furthermore, it works for pairs of Value pointers, rather than just
pairs of BasicBlock pointers.  That will be useful in the next commit,
where I want to sort values that are not basic blocks.