Pull in llvm-sra as a git submodule
llvm-sra itself is reasonably small, now that we can build it without
building an entire LLVM tree.  It really doesn't warrant a shared build
in "/p/polyglot/public/tools" as we do for giant packages like LLVM and
SAGE.  Furthermore, managing the llvm-sra build as part of
CArrayIntrospection makes it much easier for us to ensure that everyone
is using the same LLVM build tree.