Create custom "run" script for running our tool
This is similar to the "sage-opt" script from llvm-sage.  It runs
"opt" with the right environment variables set for SAGE.  Unlike
"sage-opt", though, our script always also loads our
"CArrayIntrospection.so" plugin.  Additionally, we hard-code the
absolute path to the correct "opt" so that you don't accidentally use
an incompatible LLVM build.