Extend Python path *within* SAGE-launched shell
Previously I was trying to add llvm-sage's extra search path to
$PYTHONPATH before launching "sage -sh".  But that doesn't work, because
apparently "sage -sh" completely resets $PYTHONPATH in the subshell that
it launches.

So instead, we need to modify $PYTHONPATH within that subshell.  Easily
done, since at this point we completely control the command that
subshell is asked to run.