Remove unused functionality in test code
As of right now, we're always using default JSON and ".expected" files,
never overriding any build variables other than plugin arguments, and
never running the plugin without both "-debug" and "-mem2reg".  So I've
refactored a bit to centralize repeated invariants and remove logic for
flexibility we're not actually using.  If we need more flexibility
later, we can easily recover or recreate it.