Generate Emacs customizations from build config
Some of the Emacs customizations we want involve things that depend on
the build configuration: defined macros, header search paths, and the
path to the compiler.  So this means we need to *generate* the Emacs
customization file.  We start with a generic template and then fill in
various details based on what scons has figured out.