Suppress GCC warnings in some LLVM headers
During normal compilation, we use "-isystem" instead of "-I" to add the
LLVM header search paths.  That has the effect of suppressing GCC
warnings in those LLVM headers.  Unfortunately, we don't have the
ability to use "-isystem" when Emacs Flycheck runs the compiler to check
for problems: that tool is hard-wired to use "-I".  So to avoid spurious
warning markers in my Emacs buffers, we need to take a more direct
approach to suppressing LLVM header warnings.