Avoid using "auto" with single-element init list
There's been some churn with how this is treated.  I think as of C++17
it will infer auto to be the type of the element of the list, but before
C++17 it infers auto to be the type of initializer lists.  We want the
former, but in this case the most reliable way to get that is to not use
an initializer list at all.