Add missing include of raw_ostream header
Under LLVM 3.6 and earlier, this header was indirectly being included
by some other header(s).  As of LLVM 3.7, that indirect inclusion has
gone away, so we have no choice but to include it ourselves.  We
really should have been doing that all along, but this sort of thing
is easy to overlook.