LoopInfo API updates for LLVM 3.7
In LLVM 3.6.x and earlier, LoopInfo was a function pass.  As of LLVM
3.7, LoopInfo is not a function pass, but there's a
LoopInfoWrapperPass that *is* a function pass for use with "the legacy
pass manager".  That's enough to get us building again, and in fact,
all regression tests now succeed with LLVM 3.7.0.

I want to do a bit more investigation into this "legacy pass manager"
business, though.  Perhaps there's a newer pass-wrangling scheme we
should be transitioning to instead.