Pick getelementptr type in more future-proof way
My understanding is that LLVM is moving toward a single generic
pointer type, rather than distinguishing pointer-to-int,
pointer-to-float, pointer-to-char, etc.  So one should try to avoid
relying on PointerType::getElementType().

The alternate approach I used here assumes that the new getelementptr
instruction yields the same as the existing instruction that it will
replace.  All regression tests still pass, which is encouraging.
However, based on debug output, it's not clear to me that this code is
actually covered by *any* of our tests.