Extract ValueSetSet into its own header and source
Not everyone who uses ValueSet necessarily needs to use ValueSetSet as
well.  So it's useful to be able to grab each of these without pulling
in the other.

Also, don't try to inline getValueSetFromValue.  With a nested pair of
container loops, this code is too complex to be a good inlining
candidate.

Also, pass the ValueSetSet argument to getValueSetFromValue as a const
reference, not as a copy.  There's no reason we should copy such a
complex nested container here.