Move getValueSetFromValue inside ValueSetSet
Looking at what it does, this function really does make sense as a
member function of ValueSetSet rather than a stand-alone function.
Probably the only reason it wasn't done this way before is that
ValueSetSet used to be a type alias rather than a full-fledged class.
Now that it's a real class, though, it can have members.