Run SAGE in a non-interactive bash shell
The standard SAGE script runs using "bash -i".  We don't actually need
any interactive features here, though.  Furthermore, "bash -i" emits
warning messages when run without a pseudo-tty.  That happens, for
example, during Vagrant virtual machine provisioning.
I'd rather avoid warnings, even when harmless.  The way to do that is
to not use "bash -i".

However, we do need to ensure that the "module" shell function is
defined when running the standard SAGE script.  Ordinarily that is
only defined in interactive shells, and in fact that's the whole
reason the standard SAGE script uses "bash -i" at all:
<https://bugzilla.redhat.com/show_bug.cgi?id=1238341>.  Instead of
"bash -i", we'll ensure that it's defined in our own "run" script,
then export that definition so that the SAGE script can see it too.