Promote StructElement to a full-fledged class
Previously we were using std::pair to pass these around.  But
effectively these even have member functions at this point, insofar as
there are certain functions that are always used with one of these
pairs.  I think it's time to acknowledge that StructElement is really a
class with its own behaviors, and not merely a pair of dumb data values.