Move StructElement getelementptr inspection logic
Previously this code was in a non-member getStructElement function.  Now
it's in a StructElement::get static member function.  This doesn't
really change things in any deep way, but it's more tidy to have this as
part of the StructElement class.

Incidentally, the pointer returned by this function used to leak, and
still does.  I hope to plug that leak soon with help from Valgrind and
these preliminary refactorings.