Plug FindLengthChecks leak of ValueSet instances
This approach to plugging the leak somewhat nasty.  It requires
distinguishing two different variations on ValueSetSet: one that owns
its elements and one that doesn't.  Fortunately we can make good use of
templates to avoid manually cloning the code for these two variations.
So instead of just ValueSetSet, we now have ValueSetSet<ValueSet> for
the owned case, and ValueSetSet<const ValueSet *> for the non-owned
case.  Making ValueSetSet take a template argument is a bit ugly, but we
just don't have a single uniform ownership policy across all of the
places we are using sets of sets of Value pointers.

On the positive side, our automated regression tests now produce no
Valgrind memcheck leak warnings.  Yay!