Remove several unnecessary uses of "&*"
In most normal situations, "&*p" is the same as just "p".  Certainly
that's true when "p" is a simple pointer.  These differ only when "p" is
an iterator, in which case "&*p" is an idiomatic way to get a raw
pointer to the same item that the iterator currently refers to.  But
even this is rarely needed, as iterators have overloaded operators that
let them do a pretty good job of pretending to be pointers.

Note that in two cases fixed in this commit, the redundant "&*" appeared
as part of even longer chains of pointer-fiddling operations: one was
"&**expr" and the other was "*&*expr".  By the time you have that many
"&" and "*" operations stacked up together, you really should stop and
consider whether things have gotten out of hand.