Explicitly use LLVM 3.7.0 on CS machines
Previously we were just using whatever version of LLVM corresponded to
the "/p/public/bin/llvm-config" symbolic link.  But that link could
change over time, and is not under any kind of version control.  So
using that link potentially hides important information about which
version of LLVM we prefer for a given CArrayIntrospection code snapshot.
Better to be explicit here, so that moving from one LLVM version to
another is recorded in the repository history.