Whole-file white space cleanup
I have some other changes planned for this file, but it's hard to work
in here with the indentation being so chaotic.  So as a preliminary,
separate step, I'm using "clang-format" to reindent the whole thing (and
adjust other white space too) so that it matches the style we've agreed
to use.