Improved control flow and error reporting
Previously we were catching any possible exception and assuming it
meant that the attribute we were looking for was absent.  But there's
actually an important distinction between an attribute being absent
and the same attribute being present but with a value we cannot parse
according to the expected type.  So now we only deal with the former;
exceptions due to the latter are uncaught, which at least lets us know
that the JSON we are reading is malformed.

Also, across this entire cascade of tests, use get_child_optional to
test whether a child is present or absent.  Previously we used
get_child, and let the exception reveal that something is missing.  I
find the flow using a simple series of "if ... else if ... else if"
tests more natural.

Once we determine what the annotation is for a given argument, don't
keep trying to look for alternate annotations.  Previously, even if we
saw a "sentinel" attribute, we would still look for "symbolic",
"fixed", and "other" attributes too.  That doesn't seem useful.

All regression tests still pass with these changes.