Use map::find instead of map::count + map::at
If you want to check whether a key is mapped, and then fetch the
corresponding value if there is one, use map::find.  This gives you
presence/absence information as well as access to the value in a
single lookup operation instead of two.

Same applies to unordered_map, of course.