Avoid unnecessary creation of map entries
Don't use "[]" to look up an key in a map unless you really want to
create a default-constructed entry if the key is not already present.
In this particular case, it seems we really just want the value already
associated with a key, but don't want to modify the map when the key is
absent.