Smartly use emplacement for adding defaults
map::emplace constructs a mapping entry efficiently, but does nothing if
the key already has an entry.  Either way, though, it passes back both
an iterator to the entry as well as a bool telling you whether it is a
new one or was already present.  This can help us avoid doing many
redundant look-ups, default element constructions, etc.