Adapt StructElementToValueSet to use shared_ptr
This is the first step in a long process of converting ValueSetSet to
not be a template but instead to use shared_ptr elements everywhere.
That effectively means that every ValueSet that might potentially make
it into a ValueSetSet must also be managed via shared_ptr.  In this
particular case, Annotator::runOnModule has a local ValueSetSet called
"allValueSets" into which it copies values from entries in
Annotator::structElements.  The latter has type StructElementToValueSet,
so that needs to use shared_ptr if we eventually intend for ValueSetSet
to contain shared_ptr elements.

Validated using our standard automated test suite running under Valgrind
on Fedora 23.