Don't make a copy when a pointer suffices
The container in question is never modified by Annotator.  Furthermore,
Annotator tells LLVM that it requires a FindStructElements instance, so
the copy of this container in FindStructElements will always be around
at least as long as the Annotator is using it.  Thus, there's really no
need for Annotator to keep its own private copy of this container; it
can just point at the one in FindStructElements.

(Ordinarily I would prefer to use a reference here instead of a pointer.
But Annotator::runOnModule needs to change where the pointer points.
You can do that with a non-const pointer, but you can't do that with a
reference.)

Validated using our standard automated test suite running under Valgrind
on Fedora 23.