Add ValueSetSet shared_ptr specialization
We're not actually using this yet anywhere, but I think we may soon.
The plan is that eventually the other two specializations will no longer
be used anywhere, at which point ValueSetSet can be demoted from a
template to a regular class that always uses shared_ptr for its
elements.