White-space cleanups before larger changes
This reformatting was done using "clang-format", targeting the
indentation style we've already agreed to use as recorded in
".clang-format".

I don't usually like to do big white-space changes like this, as it
tends to create lots of merge conflicts later.  But the indentation in
the file right now is quite chaotic, which makes it difficult to apply
the more substantive changes I'm working on.