Prefer emplacement over "[]" for novel insertions
Also add an "assert" that represents my belief that this is indeed a
novel insertion.  I don't *think* there should already be an item in
this map for this key, and if I'm wrong, I want to know about it.  (It
may not actually matter if I'm wrong in this specific case, but I'd
still like to know.)